requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
PyGithub==2.3.0
httpx==0.27.2
pydantic==2.8.2
pygments==2.18.0
unidiff==0.7.5
langchain==0.3.14
langchain-openai==0.3.1
langchain-core==0.3.29
pydantic==2.8.2
openai==1.59.6
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.3.2
sqlalchemy==2.0.36
sentence-transformers==3.3.1
tiktoken==0.8.0
gitpython==3.1.43